requirements.txt
# Core dependencies
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10

# Document processing
pypdf>=3.17.0
python-docx>=1.0.0
openpyxl>=3.1.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0